+++ index.html
@@ -234,9 +234,10 @@
       <h2 class="data-error__title">Не удалось загрузить данные</h2>
     </section>
   </template>
-  <script src="/js/functions.js"></script>
-  <script src="/js/main.js"></script>
-  <script  type = "module" src="/js/main.js"></script>
+  <script src="/js/dz-2-functions.js"></script>
+  <script src="/js/dz4-main.js"></script>
+  <script type="module" src="/js/main.js"></script>
+  <script>src = "functions.js"</script>
 </body>
 
 </html>
