+++ js/data.js
@@ -1,0 +1,52 @@
+/* eslint-disable no-unused-vars */
+/* eslint-disable no-console */
+
+const DESCRIPTTIONS = () => ([
+  'Леонардо да Винчи, «Мона Лиза».',
+  'Анри Матисс, «Танец».',
+  'Микеланджело Буонарроти, «Сотворение Адама».',
+  'Гюстав Курбе, «Отчаяние. Автопортрет».',
+  'Карл Брюллов, «Всадница».',
+  'Пьер Огюст Ренуар, «Завтрак гребцов».',
+  'Рафаэль Санти, «Сикстинская Мадонна».',
+  'Иероним Босх, «Сад земных наслаждений».',
+  'Диего Веласкес, «Менины».',
+  'Эжен Делакруа, «Свобода, ведущая народ».',
+  'Михаил Врубель, «Демон сидящий».',
+  'Грант Вуд, «Американская готика».',
+  'Рене Магритт, «Сын человеческий».',
+  'Сальвадор Дали, «Постоянство памяти».',
+  'Иван Айвазовский, «Девятый вал».',
+  'Питер Брейгель, «Вавилонская башня».',
+  'Леонардо да Винчи, «Тайная вечеря».',
+  'Эдуард Мане, «Бар в „Фоли-Бержер“».',
+  'Эдгар Дега, «Голубые танцовщицы».',
+  'Рембрандт, «Ночной дозор».',
+  'Эдвард Мунк, «Крик».',
+  'Жак-Луи Давид, «Смерть Марата».',
+  'Винсент Ван Гог, «Звёздная ночь».',
+  'Сандро Боттичелли, «Рождение Венеры».',
+  'Ян Вермеер, «Девушка с жемчужной серёжкой».',
+]);
+
+const MESSAGES = () => ([
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+]);
+
+const NAMES = () => ([
+  'Даниил',
+  'Тимофей',
+  'Владислав',
+  'Игорь',
+  'Владимир',
+  'Руслан',
+]);
+
+// const getValue = () => [DESCRIPTTIONS(), MESSAGES(), NAMES()];
+
+export {DESCRIPTTIONS, MESSAGES, NAMES};
