+++ js/getComments.js
@@ -1,0 +1,14 @@
+import {MESSAGES, NAMES} from './data.js';
+import {getRandomArrayElement, getGeneratorRandomInteger} from './getRandomArrayElement.js';
+import {getCounter} from './getCounter.js';
+
+const getCounterCommentId = getCounter();
+
+const getComments = () => ({
+  id: getCounterCommentId(),// не должны повторяться
+  avatar: `img/avatar-${getGeneratorRandomInteger(1, 6)}.svg`,
+  message: `${getRandomArrayElement(MESSAGES)}`,//все случайные,могут повторяться
+  name: `${getRandomArrayElement(NAMES)}`,//все случайные,могут повторяться
+});
+
+export {getComments};
