+++ js/getCounter.js
@@ -1,0 +1,6 @@
+const getCounter = () => {
+  let firstCounter = 0;
+  return () => (firstCounter += 1);
+};
+
+export {getCounter};
